Change names to be consistently pluralised
