+++ lib/index.js
@@ -1,14 +1,14 @@
 module.exports = [
   require("./rules/alpha-values"),
-  require("./rules/border-width"),
-  require("./rules/font-family"),
-  require("./rules/font-size"),
-  require("./rules/font-weight"),
-  require("./rules/letter-spacing"),
-  require("./rules/line-height"),
+  require("./rules/border-widths"),
+  require("./rules/font-families"),
+  require("./rules/font-sizes"),
+  require("./rules/font-weights"),
+  require("./rules/letter-spacings"),
+  require("./rules/line-heights"),
   require("./rules/radii"),
   require("./rules/sizes"),
   require("./rules/space"),
-  require("./rules/word-spacing"),
+  require("./rules/word-spacings"),
   require("./rules/z-indices"),
 ];
