+++ lib/rules/border-widths/index.js
@@ -12,10 +12,10 @@ const hasScalesWithUnits = require("../../utils/hasScalesWithUnits");
 const isOnNumericScale = require("../../utils/isOnNumericScale");
 const setValue = require("../../utils/setValue");
 
-const ruleName = "scales/letter-spacing";
+const ruleName = "scales/border-widths";
 const messages = createRuleMessages(ruleName);
 
-const propertyFilter = "letter-spacing";
+const propertyFilter = /^border$|^border.*(width$|top$|right$|bottom$|left$|block$|inline$|start$|end$)/;
 
 const rule = (primary, secondary, { fix }) => {
   return (root, result) => {
