+++ lib/rules/font-families/index.js
@@ -10,7 +10,7 @@ const { parse: parseFontFamily } = require("font-family");
 const createRuleMessages = require("../../utils/createRuleMessages");
 const parseShorthandFont = require("../../utils/parseShorthandFont");
 
-const ruleName = "scales/font-family";
+const ruleName = "scales/font-families";
 const messages = createRuleMessages(ruleName);
 
 // Rule accepts an array of numbers for the scale
