+++ lib/rules/font-sizes/index.js
@@ -9,13 +9,14 @@ const findScaleByUnit = require("../../utils/findScaleByUnit");
 const getClosest = require("../../utils/getClosest");
 const getValue = require("../../utils/getValue");
 const hasScalesWithUnits = require("../../utils/hasScalesWithUnits");
+const isLineHeight = require("../../utils/isLineHeight");
 const isOnNumericScale = require("../../utils/isOnNumericScale");
 const setValue = require("../../utils/setValue");
 
-const ruleName = "scales/border-width";
+const ruleName = "scales/font-sizes";
 const messages = createRuleMessages(ruleName);
 
-const propertyFilter = /^border$|^border.*(width$|top$|right$|bottom$|left$|block$|inline$|start$|end$)/;
+const propertyFilter = /^font-size$|^font$/;
 
 const rule = (primary, secondary, { fix }) => {
   return (root, result) => {
@@ -28,6 +29,7 @@ const rule = (primary, secondary, { fix }) => {
       return;
 
     root.walkDecls(propertyFilter, (decl) => {
+      const { prop } = decl;
       const value = getValue(decl);
       let hasFix = false;
 
@@ -35,9 +37,18 @@ const rule = (primary, secondary, { fix }) => {
         ignoreUnknownWords: true,
       });
 
-      valueRoot.walkNumerics((node) => {
-        check(node);
-      });
+      switch (prop) {
+        case "font": {
+          const node = findFontSize(valueRoot.nodes);
+          if (node) check(node);
+          break;
+        }
+        case "font-size":
+          valueRoot.walkNumerics((node) => {
+            check(node);
+          });
+          break;
+      }
 
       function check(node) {
         const { value, unit } = node;
@@ -65,6 +76,14 @@ const rule = (primary, secondary, { fix }) => {
     });
   };
 };
+
+function findFontSize(nodes) {
+  const node = nodes.find(
+    ({ type, unit }) => type === "numeric" && unit !== ""
+  );
+
+  if (node && !isLineHeight(node)) return node;
+}
 
 rule.primaryOptionArray = true;
