+++ lib/rules/font-weights/index.js
@@ -11,7 +11,7 @@ const isLineHeight = require("../../utils/isLineHeight");
 const isOnNumericScale = require("../../utils/isOnNumericScale");
 const setValue = require("../../utils/setValue");
 
-const ruleName = "scales/font-weight";
+const ruleName = "scales/font-weights";
 const messages = createRuleMessages(ruleName);
 
 const propertyFilter = /^font-weight$|^font$/;
