+++ lib/rules/letter-spacings/README.md
@@ -1,18 +1,18 @@
-# word-spacing
+# letter-spacings
 
-Specify scales for word-spacing.
+Specify scales for letter spacing.
 
 ```css
 a {
-  word-spacing: 0.1rem;
+  letter-spacing: 0.1rem;
 }
 /**               ↑
- *                This word spacing */
+ *                This letter spacing */
 ```
 
 This rule can automatically fix all of the problems reported.
 
-This rule checks `<length>` and `<percentage>` values.
+This rule checks `<length>` values.
 
 This rule can be paired with the [`declaration-property-unit-allowed-list`](https://stylelint.io/user-guide/rules/declaration-property-unit-allowed-list) rule in stylelint.
 
@@ -23,25 +23,17 @@ This rule can be paired with the [`declaration-property-unit-allowed-list`](http
 Given:
 
 ```json
-[
-  {
-    "scale": [-0.1, 0.2],
-    "units": ["rem"]
-  }
-]
+{
+  "scale": [0.1, 0.2],
+  "units": ["rem"]
+}
 ```
 
 The following patterns are considered violations:
 
 ```css
 a {
-  word-spacing: 0.1rem;
-}
-```
-
-```css
-a {
-  word-spacing: 0.5rem;
+  letter-spacing: 0.5rem;
 }
 ```
 
@@ -49,12 +41,10 @@ The following patterns are _not_ considered violations:
 
 ```css
 a {
-  word-spacing: -0.1rem;
+  letter-spacing: 0.1rem;
 }
-```
 
-```css
 a {
-  word-spacing: 0.2rem;
+  letter-spacing: 0.2rem;
 }
 ```
