+++ lib/rules/letter-spacings/__tests__/index.js
