+++ lib/rules/letter-spacings/index.js
@@ -12,10 +12,10 @@ const hasScalesWithUnits = require("../../utils/hasScalesWithUnits");
 const isOnNumericScale = require("../../utils/isOnNumericScale");
 const setValue = require("../../utils/setValue");
 
-const ruleName = "scales/word-spacing";
+const ruleName = "scales/letter-spacings";
 const messages = createRuleMessages(ruleName);
 
-const propertyFilter = "word-spacing";
+const propertyFilter = "letter-spacing";
 
 const rule = (primary, secondary, { fix }) => {
   return (root, result) => {
