+++ lib/rules/line-heights/README.md
@@ -1,6 +1,6 @@
-# line-height
+# line-heights
 
-Specify a scale for line-heights.
+Specify a scale for line heights.
 
 ```css
 a {
@@ -12,7 +12,7 @@ a {
 
 This rule can automatically fix all of the problems reported.
 
-This rule checks unitless line-heights.
+This rule checks unitless line heights.
 
 This rule can be paired with the [`declaration-property-unit-allowed-list`](https://stylelint.io/user-guide/rules/declaration-property-unit-allowed-list) rule in stylelint.
