+++ lib/rules/line-heights/index.js
@@ -5,25 +5,22 @@ const {
 } = require("stylelint");
 
 const createRuleMessages = require("../../utils/createRuleMessages");
-const findScaleByUnit = require("../../utils/findScaleByUnit");
 const getClosest = require("../../utils/getClosest");
 const getValue = require("../../utils/getValue");
-const hasScalesWithUnits = require("../../utils/hasScalesWithUnits");
-const isLineHeight = require("../../utils/isLineHeight");
 const isOnNumericScale = require("../../utils/isOnNumericScale");
 const setValue = require("../../utils/setValue");
 
-const ruleName = "scales/font-size";
+const ruleName = "scales/line-heights";
 const messages = createRuleMessages(ruleName);
 
-const propertyFilter = /^font-size$|^font$/;
+const propertyFilter = /^line-height$|^font$/;
 
 const rule = (primary, secondary, { fix }) => {
   return (root, result) => {
     if (
       !validateOptions(result, ruleName, {
         actual: primary,
-        possible: hasScalesWithUnits,
+        possible: Array.isArray,
       })
     )
       return;
@@ -39,11 +36,11 @@ const rule = (primary, secondary, { fix }) => {
 
       switch (prop) {
         case "font": {
-          const node = findFontSize(valueRoot.nodes);
+          const node = findLineHeight(valueRoot.nodes);
           if (node) check(node);
           break;
         }
-        case "font-size":
+        case "line-height":
           valueRoot.walkNumerics((node) => {
             check(node);
           });
@@ -53,18 +50,18 @@ const rule = (primary, secondary, { fix }) => {
       function check(node) {
         const { value, unit } = node;
 
-        const scale = findScaleByUnit(primary, unit);
+        if (unit) return;
 
-        if (isOnNumericScale(scale, value)) return;
+        if (isOnNumericScale(primary, value)) return;
 
         if (fix) {
-          node.value = getClosest(scale, value);
+          node.value = getClosest(primary, value);
           hasFix = true;
           return;
         }
 
         report({
-          message: messages.expected(value, scale.join(", "), unit),
+          message: messages.expected(value, primary.join(", ")),
           node: decl,
           result,
           ruleName,
@@ -77,12 +74,11 @@ const rule = (primary, secondary, { fix }) => {
   };
 };
 
-function findFontSize(nodes) {
+function findLineHeight(nodes) {
   const node = nodes.find(
-    ({ type, unit }) => type === "numeric" && unit !== ""
+    ({ type, value }) => type === "operator" && value === "/"
   );
-
-  if (node && !isLineHeight(node)) return node;
+  if (node) return node.next();
 }
 
 rule.primaryOptionArray = true;
