+++ lib/rules/word-spacings/README.md
@@ -1,18 +1,18 @@
-# letter-spacing
+# word-spacings
 
-Specify scales for letter-spacing.
+Specify scales for word spacings.
 
 ```css
 a {
-  letter-spacing: 0.1rem;
+  word-spacing: 0.1rem;
 }
 /**               ↑
- *                This letter spacing */
+ *                This word spacing */
 ```
 
 This rule can automatically fix all of the problems reported.
 
-This rule checks `<length>` values.
+This rule checks `<length>` and `<percentage>` values.
 
 This rule can be paired with the [`declaration-property-unit-allowed-list`](https://stylelint.io/user-guide/rules/declaration-property-unit-allowed-list) rule in stylelint.
 
@@ -23,17 +23,25 @@ This rule can be paired with the [`declaration-property-unit-allowed-list`](http
 Given:
 
 ```json
-{
-  "scale": [0.1, 0.2],
-  "units": ["rem"]
-}
+[
+  {
+    "scale": [-0.1, 0.2],
+    "units": ["rem"]
+  }
+]
 ```
 
 The following patterns are considered violations:
 
 ```css
 a {
-  letter-spacing: 0.5rem;
+  word-spacing: 0.1rem;
+}
+```
+
+```css
+a {
+  word-spacing: 0.5rem;
 }
 ```
 
@@ -41,10 +49,12 @@ The following patterns are _not_ considered violations:
 
 ```css
 a {
-  letter-spacing: 0.1rem;
+  word-spacing: -0.1rem;
 }
+```
 
+```css
 a {
-  letter-spacing: 0.2rem;
+  word-spacing: 0.2rem;
 }
 ```
