+++ lib/rules/word-spacings/__tests__/index.js
