+++ lib/rules/word-spacings/index.js
@@ -5,28 +5,29 @@ const {
 } = require("stylelint");
 
 const createRuleMessages = require("../../utils/createRuleMessages");
+const findScaleByUnit = require("../../utils/findScaleByUnit");
 const getClosest = require("../../utils/getClosest");
 const getValue = require("../../utils/getValue");
+const hasScalesWithUnits = require("../../utils/hasScalesWithUnits");
 const isOnNumericScale = require("../../utils/isOnNumericScale");
 const setValue = require("../../utils/setValue");
 
-const ruleName = "scales/line-height";
+const ruleName = "scales/word-spacings";
 const messages = createRuleMessages(ruleName);
 
-const propertyFilter = /^line-height$|^font$/;
+const propertyFilter = "word-spacing";
 
 const rule = (primary, secondary, { fix }) => {
   return (root, result) => {
     if (
       !validateOptions(result, ruleName, {
         actual: primary,
-        possible: Array.isArray,
+        possible: hasScalesWithUnits,
       })
     )
       return;
 
     root.walkDecls(propertyFilter, (decl) => {
-      const { prop } = decl;
       const value = getValue(decl);
       let hasFix = false;
 
@@ -34,34 +35,25 @@ const rule = (primary, secondary, { fix }) => {
         ignoreUnknownWords: true,
       });
 
-      switch (prop) {
-        case "font": {
-          const node = findLineHeight(valueRoot.nodes);
-          if (node) check(node);
-          break;
-        }
-        case "line-height":
-          valueRoot.walkNumerics((node) => {
-            check(node);
-          });
-          break;
-      }
+      valueRoot.walkNumerics((node) => {
+        check(node);
+      });
 
       function check(node) {
         const { value, unit } = node;
 
-        if (unit) return;
+        const scale = findScaleByUnit(primary, unit);
 
-        if (isOnNumericScale(primary, value)) return;
+        if (isOnNumericScale(scale, value)) return;
 
         if (fix) {
-          node.value = getClosest(primary, value);
+          node.value = getClosest(scale, value);
           hasFix = true;
           return;
         }
 
         report({
-          message: messages.expected(value, primary.join(", ")),
+          message: messages.expected(value, scale.join(", "), unit),
           node: decl,
           result,
           ruleName,
@@ -73,13 +65,6 @@ const rule = (primary, secondary, { fix }) => {
     });
   };
 };
-
-function findLineHeight(nodes) {
-  const node = nodes.find(
-    ({ type, value }) => type === "operator" && value === "/"
-  );
-  if (node) return node.next();
-}
 
 rule.primaryOptionArray = true;
